remove second restaurant vc & add menu buttons
